refactor(client): load only the selected backend graph
