+++ src/notebooklm/raw.py
@@ -15,15 +15,18 @@ compatibility surface of this package.
 from __future__ import annotations
 
 import asyncio
+import importlib
 import re
-from collections.abc import AsyncIterator, Callable, Sequence
+from collections.abc import Callable, Sequence
 from dataclasses import dataclass
 from enum import Enum
-from typing import Any, Generic, Protocol, TypeVar, cast
+from typing import TYPE_CHECKING, Any, Generic, Protocol, TypeVar, cast
 
-from ._android.errors import sanitize_async_boundary, sanitize_escaping_exception
-from ._web.raw import WebRawAPI
 from .exceptions import DecodingError
+
+if TYPE_CHECKING:
+    from ._android.raw import AndroidRawAPI
+    from ._web.raw import WebRawAPI
 
 __all__ = [
     "AndroidRawAPI",
@@ -159,6 +162,8 @@ def _bytes_size(payload: bytes) -> int:
 
 
 def _codec_failure(*, path: str, stage: str, error: BaseException) -> DecodingError:
+    from ._android.errors import sanitize_escaping_exception
+
     sanitize_escaping_exception(error)
     return DecodingError(
         f"Raw Android gRPC {stage} failed.",
@@ -226,126 +231,19 @@ def _validated_metadata(metadata: GrpcMetadata | None) -> tuple[tuple[str, str |
     return tuple(validated)
 
 
-class AndroidRawAPI:
-    """Raw unary and unary-stream Android calls over one ``AndroidSession``."""
+_LAZY_EXPORTS = {
+    "AndroidRawAPI": ("notebooklm._android.raw", "AndroidRawAPI"),
+    "WebRawAPI": ("notebooklm._web.raw", "WebRawAPI"),
+}
 
-    def __init__(self, session: Any) -> None:
-        self._transport = session
 
-    @sanitize_async_boundary
-    async def unary(
-        self,
-        method: GrpcUnaryMethod[RequestT, ResponseT],
-        request: RequestT,
-        *,
-        timeout: float | None = None,
-        metadata: GrpcMetadata | None = None,
-    ) -> ResponseT:
-        """Invoke one raw unary method with an explicit replay classification."""
+def __getattr__(name: str) -> object:
+    target = _LAZY_EXPORTS.get(name)
+    if target is None:
+        raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+    module_name, attribute = target
+    return getattr(importlib.import_module(module_name), attribute)
 
-        # Keep caller-controlled codecs outside grpcio.  grpcio may otherwise
-        # collapse a serializer/deserializer exception into an opaque UNKNOWN
-        # status before the library can sanitize it.
-        safe_metadata = _validated_metadata(metadata)
-        wire_request = _serialize_request(method, request)
-        replay_safe = method.replay_policy is ReplayPolicy.SAFE_READ
 
-        from ._android.session import classify_raw_replay
-
-        wire_response = await self._transport.unary(
-            method.path,
-            wire_request,
-            replay_safe=replay_safe,
-            timeout=timeout,
-            response_type=None,
-            telemetry_method="raw.unary",
-            metadata=safe_metadata,
-            request_serializer=_identity_bytes,
-            response_deserializer=_identity_bytes,
-            raw_replay=classify_raw_replay(replay_safe),
-        )
-        return _deserialize_response(method, wire_response)
-
-    def unary_stream(
-        self,
-        method: GrpcUnaryStreamMethod[RequestT, ResponseT],
-        request: RequestT,
-        *,
-        timeout: float | None = None,
-        max_response_bytes: int | None = _DEFAULT_RAW_STREAM_MAX_RESPONSE_BYTES,
-        metadata: GrpcMetadata | None = None,
-    ) -> AsyncIterator[ResponseT]:
-        """Yield one bounded raw server stream; streams are never replayed."""
-
-        return self._unary_stream_impl(
-            method,
-            request,
-            timeout=timeout,
-            max_response_bytes=max_response_bytes,
-            metadata=metadata,
-        )
-
-    async def _unary_stream_impl(
-        self,
-        method: GrpcUnaryStreamMethod[RequestT, ResponseT],
-        request: RequestT,
-        *,
-        timeout: float | None,
-        max_response_bytes: int | None,
-        metadata: GrpcMetadata | None,
-    ) -> AsyncIterator[ResponseT]:
-        session = self._transport
-        iterator: AsyncIterator[bytes] | None = None
-        failure: BaseException | None = None
-        safe_metadata: tuple[tuple[str, str | bytes], ...] | None = None
-        wire_request: bytes | None = None
-        wire_response: bytes | None = None
-        try:
-            if max_response_bytes is not None and max_response_bytes < 1:
-                raise ValueError("max_response_bytes must be >= 1 when supplied")
-            safe_metadata = _validated_metadata(metadata)
-            wire_request = _serialize_request(method, request)
-            replay_safe = method.replay_policy is ReplayPolicy.SAFE_READ
-
-            from ._android.session import classify_raw_replay
-
-            iterator = session.stream(
-                method.path,
-                wire_request,
-                replay_safe=replay_safe,
-                timeout=timeout,
-                response_type=None,
-                telemetry_method="raw.unary_stream",
-                max_response_bytes=max_response_bytes,
-                metadata=safe_metadata,
-                request_serializer=_identity_bytes,
-                response_deserializer=_identity_bytes,
-                response_sizer=_bytes_size,
-                raw_replay=classify_raw_replay(replay_safe),
-            )
-            async for wire_response in iterator:
-                yield _deserialize_response(method, wire_response)
-        except BaseException as error:
-            failure = sanitize_escaping_exception(error)
-        finally:
-            if iterator is not None:
-                try:
-                    close = getattr(iterator, "aclose", None)
-                    if close is not None:
-                        await close()
-                except BaseException as error:
-                    if failure is None:
-                        failure = sanitize_escaping_exception(error)
-            del (
-                self,
-                session,
-                iterator,
-                request,
-                method,
-                metadata,
-                safe_metadata,
-                wire_request,
-                wire_response,
-            )
-        if failure is not None:
-            raise failure from None
+def __dir__() -> list[str]:
+    return sorted({*globals(), *__all__})
